examples/babylon-langgraph-agent/pyproject.toml
[project]
name = "babylon-langgraph-agent"
version = "1.0.0"
description = "Autonomous Babylon trading agent with LangGraph and Agent0"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "langgraph>=0.2.0",
    "a2a-sdk>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "web3>=7.0.0",
    "eth-account>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-xprocess>=1.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

